Move PaddedGraphSequence to be adjacent to its generator (#1607)
It is a bit clearer to group each generator with its corresponding sequence,
rather than grouping all the sequences.

This is helpful for #1523.
