+++ stellargraph/mapper/padded_graph_generator.py
@@ -14,9 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from ..core.graph import StellarGraph
-from ..core.utils import is_real_iterable
-from .sequences import PaddedGraphSequence
+from ..core.utils import is_real_iterable, normalize_adj
+from ..random import random_state
 import numpy as np
+from tensorflow.keras.utils import Sequence
 from .base import Generator
 
 
@@ -149,3 +150,134 @@ class PaddedGraphGenerator(Generator):
             shuffle=shuffle,
             seed=seed,
         )
+
+
+class PaddedGraphSequence(Sequence):
+    """
+    A Keras-compatible data generator for training and evaluating graph classification models.
+    Use this class with the Keras methods :meth:`keras.Model.fit`,
+        :meth:`keras.Model.evaluate`, and
+        :meth:`keras.Model.predict`,
+
+    This class should be created using the `.flow(...)` method of
+    :class:`PaddedGraphGenerator`.
+
+    Args:
+        graphs (list)): The graphs as StellarGraph objects.
+        targets (np.ndarray, optional): An optional array of graph targets of size (N x C),
+            where N is the number of graphs and C is the target size (e.g., number of classes.)
+        normalize (bool, optional): Specifies whether the adjacency matrix for each graph should
+            be normalized or not. The default is True.
+        symmetric_normalization (bool, optional): Use symmetric normalization if True, that is left and right multiply
+            the adjacency matrix by the inverse square root of the degree matrix; otherwise left multiply the adjacency
+            matrix by the inverse of the degree matrix. This parameter is ignored if normalize=False.
+        batch_size (int, optional): The batch size. It defaults to 1.
+        name (str, optional): An optional name for this generator object.
+        shuffle (bool, optional): If True the node IDs will be shuffled at the end of each epoch.
+        seed (int, optional): Random seed.
+    """
+
+    def __init__(
+        self,
+        graphs,
+        targets=None,
+        normalize=True,
+        symmetric_normalization=True,
+        batch_size=1,
+        name=None,
+        shuffle=False,
+        seed=None,
+    ):
+
+        self.name = name
+        self.graphs = np.asanyarray(graphs)
+        self.normalize_adj = normalize
+        self.targets = targets
+        self.batch_size = batch_size
+
+        if targets is not None:
+            if len(graphs) != len(targets):
+                raise ValueError(
+                    "expected the number of target values and the number of graphs to be the same length,"
+                    f"found {len(graphs)} graphs and {len(targets)} targets."
+                )
+
+            self.targets = np.asanyarray(targets)
+
+        if self.normalize_adj:
+            self.normalized_adjs = [
+                normalize_adj(
+                    graph.to_adjacency_matrix(),
+                    symmetric=symmetric_normalization,
+                    add_self_loops=True,
+                )
+                for graph in graphs
+            ]
+        else:
+            self.normalize_adjs = [graph.to_adjacency_matrix() for graph in graphs]
+
+        self.normalized_adjs = np.asanyarray(self.normalized_adjs)
+        _, self._np_rs = random_state(seed)
+        self.shuffle = shuffle
+
+        self.on_epoch_end()
+
+    def __len__(self):
+        return int(np.ceil(len(self.graphs) / self.batch_size))
+
+    def __getitem__(self, index):
+
+        batch_start, batch_end = index * self.batch_size, (index + 1) * self.batch_size
+
+        graphs = self.graphs[batch_start:batch_end]
+        adj_graphs = self.normalized_adjs[batch_start:batch_end]
+
+        # The number of nodes for the largest graph in the batch. We are going to pad with 0 rows and columns
+        # the adjacency and node feature matrices (only the rows in this case) to equal in size the adjacency and
+        # feature matrices of the largest graph.
+        max_nodes = max([graph.number_of_nodes() for graph in graphs])
+
+        graph_targets = None
+        if self.targets is not None:
+            graph_targets = self.targets[batch_start:batch_end]
+
+        # pad adjacency and feature matrices to equal the size of those from the largest graph
+        features = [
+            np.pad(
+                graph.node_features(),
+                pad_width=((0, max_nodes - graph.number_of_nodes()), (0, 0)),
+            )
+            for graph in graphs
+        ]
+        features = np.stack(features)
+
+        for adj in adj_graphs:
+            adj.resize((max_nodes, max_nodes))
+        adj_graphs = np.stack([adj.toarray() for adj in adj_graphs])
+
+        masks = np.full((len(graphs), max_nodes), fill_value=False, dtype=np.bool)
+        for index, graph in enumerate(graphs):
+            masks[index, : graph.number_of_nodes()] = True
+
+        # features is array of dimensionality
+        #      batch size x N x F
+        # masks is array of dimensionality
+        #      batch size x N
+        # adj_graphs is array of dimensionality
+        #      batch size x N x N
+        # graph_targets is array of dimensionality
+        #      batch size x C
+        # where N is the maximum number of nodes for largest graph in the batch, F is
+        # the node feature dimensionality, and C is the number of target classes
+        return [features, masks, adj_graphs], graph_targets
+
+    def on_epoch_end(self):
+        """
+         Shuffle all graphs at the end of each epoch
+        """
+        if self.shuffle:
+            indexes = self._np_rs.permutation(len(self.graphs))
+            self.graphs = self.graphs[indexes]
+            self.normalized_adjs = self.normalized_adjs[indexes]
+            if self.targets is not None:
+                self.targets = self.targets[indexes]
